chore(api): move tests into right context
