+++ api/tests/prescription/organization-learner/unit/domain/models/ParticipantRepartition_test.js
@@ -1,5 +1,5 @@
-import { ParticipantRepartition } from '../../../../src/prescription/organization-learner/domain/models/ParticipantRepartition.js';
-import { expect } from '../../../test-helper.js';
+import { ParticipantRepartition } from '../../../../../../src/prescription/organization-learner/domain/models/ParticipantRepartition.js';
+import { expect } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Models | ParticipantRepartition', function () {
   it('counts the number of registred users', function () {
