+++ api/tests/prescription/target-profile/unit/domain/models/TargetProfileSummaryForAdmin_test.js
@@ -1,4 +1,4 @@
-import { domainBuilder, expect } from '../../../test-helper.js';
+import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Models | TargetProfileSummaryForAdmin', function () {
   describe('#canDetach', function () {
