+++ api/tests/prescription/target-profile/unit/domain/models/TargetProfile_test.js
@@ -1,5 +1,5 @@
 import { TargetProfile } from '../../../../../../src/prescription/target-profile/domain/models/TargetProfile.js';
-import { expect } from '../../../../../test-helper.js';
+import { domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Unit | Domain | Models | TargetProfile', function () {
   describe('#detach', function () {
@@ -17,6 +17,25 @@ describe('Unit | Domain | Models | TargetProfile', function () {
       targetProfile.detach([3, 3]);
 
       expect(targetProfile.organizationIdsToDetach).deep.equal([3]);
+    });
+  });
+
+  describe('#hasBadges', function () {
+    it('should return true when target profile has badges', function () {
+      // given
+      const badge = domainBuilder.buildBadge();
+      const targetProfile = domainBuilder.buildTargetProfile({ badges: [badge] });
+
+      // then
+      expect(targetProfile.hasBadges).to.be.true;
+    });
+
+    it("should return false when target profile doesn't have badges", function () {
+      // given
+      const targetProfile = domainBuilder.buildTargetProfile();
+
+      // then
+      expect(targetProfile.hasBadges).to.be.false;
     });
   });
 });
